--- flute/csrc/hadamard.h
@@ -1,23 +1,0 @@
-// Based on code by Tri Dao, 2023
-
-#pragma once
-
-#include <torch/extension.h>
-
-
-struct HadamardParamsBase {
-    using index_t = int64_t;
-
-    int batch, dim, log_N;
-
-    index_t x_batch_stride;
-    index_t out_batch_stride;
-
-    float scale;
-
-    // Common data pointers.
-    void *__restrict__ x_ptr;
-    void *__restrict__ out_ptr;
-};
-
-at::Tensor fast_hadamard_transform(at::Tensor &x, float scale);
