+++ setup.py
@@ -93,8 +93,7 @@ def get_extensions() -> List:
     fht_dir = os.path.join(this_dir, "fast-hadamard-transform", "csrc")
     include_dirs.append(fht_dir)
     sources += (
-        list(glob.glob(os.path.join(fht_dir, "*.cu"))) +
-        list(glob.glob(os.path.join(fht_dir, "*.cpp"))))
+        list(glob.glob(os.path.join(fht_dir, "*.cu"))))
 
     ext_modules = [
         CUDAExtension(
